Start and stop activities in Timeular on changing the ZEI's upside
This is a basic API implementation which naivly starts and stops activities.
Whenever a change of sides is detected, the current tracking object is retrieved and compared with the one related
to the device side to support handling the cases where the user manually changed the current tracking in the app.

The list of activities are only loaded on program initialization, so there is no support yet for reloading the activity
details on change.