Do not crash when launched with a non-activity side
When the application is launched and connected with the Zei, but the Zei is positioned on
a side without an activity registered, there is nothing to pass to the "CurrentTracking"
state. This caused the application to end with a `nil pointer dereference`

Test Plan: Build and run the application with the Zei positioned in a non-activity side up.
Launch it again, but now with an activity side up. In both cases, see that the app correctly
synchronizes state